bake: Enable multiprocess baking.
Baking is now done by running a worker per CPU, and sending jobs to them.
This changes several things across the codebase:

* Ability to not cache things related to pages other than the 'main' page
  (i.e. the page at the bottom of the execution stack).
* Decouple the baking process from the bake records, so only the main process
  keeps track (and modifies) the bake record.
* Remove the need for 'batch page getters' and loading a page directly from
  the page factories.

There are various smaller changes too included here, including support for
scope performance timers that are saved with the bake record and can be
printed out to the console. Yes I got carried away.

For testing, the in-memory 'mock' file-system doesn't work anymore, since
we're spawning processes, so this is replaced by a 'tmpfs' file-system which
is saved in temporary files on disk and deleted after tests have run.